Fixed iOS9.1 Warnings and Errors

+++ Lib/Azure Storage Client Library/Azure Storage Client LibraryTests/AZSBlobSASTests.m
@@ -67,7 +67,7 @@
 - (NSString *) URLEncodedStringWithString:(NSString *)stringToConvert {
     NSMutableString * output = [NSMutableString string];
     const unsigned char * source = (const unsigned char *)[stringToConvert UTF8String];
-    int sourceLen = strlen((const char *)source);
+    unsigned long sourceLen = strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
